Move tests from mailqueue.py into test_mailqueue.py
